renames SprkRadio to SprkRadioItem and groups the components into a folder
