+++ react/src/base/inputs/SprkRadio/SprkRadioItem/SprkRadioItem.test.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import Enzyme, { mount, shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import SprkRadio from './SprkRadio';
+import SprkRadioItem from './SprkRadioItem';
 
 Enzyme.configure({ adapter: new Adapter() });
 
 describe('SprkRadio:', () => {
   it('should apply additional classes', () => {
     const expected = 'test-class';
-    const wrapper = shallow(<SprkRadio additionalClasses={expected} />);
+    const wrapper = shallow(<SprkRadioItem additionalClasses={expected} />);
     expect(wrapper.find('.sprk-b-SelectionContainer').hasClass(expected)).toBe(
       true,
     );
@@ -16,25 +16,29 @@ describe('SprkRadio:', () => {
 
   it('should apply additional radio classes', () => {
     const expected = 'test-class';
-    const wrapper = shallow(<SprkRadio radioAdditionalClasses={expected} />);
+    const wrapper = shallow(
+      <SprkRadioItem radioAdditionalClasses={expected} />,
+    );
     expect(wrapper.find('.sprk-b-Radio__input').hasClass(expected)).toBe(true);
   });
 
   it('should apply additional label classes', () => {
     const expected = 'test-class';
-    const wrapper = shallow(<SprkRadio labelAdditionalClasses={expected} />);
+    const wrapper = shallow(
+      <SprkRadioItem labelAdditionalClasses={expected} />,
+    );
     expect(wrapper.find('.sprk-b-Radio__label').hasClass(expected)).toBe(true);
   });
 
   it('should apply huge container class', () => {
-    const wrapper = shallow(<SprkRadio variant="huge" />);
+    const wrapper = shallow(<SprkRadioItem variant="huge" />);
     expect(
       wrapper.find('.sprk-b-SelectionContainer').hasClass('sprk-b-Radio--huge'),
     ).toBe(true);
   });
 
   it('should render label', () => {
-    const wrapper = shallow(<SprkRadio>Label</SprkRadio>);
+    const wrapper = shallow(<SprkRadioItem>Label</SprkRadioItem>);
     expect(wrapper.find('label').text()).toEqual('Label');
     expect(
       wrapper
@@ -45,7 +49,7 @@ describe('SprkRadio:', () => {
 
   it('should apply ariaDescribedBy', () => {
     const expected = 'test-aria';
-    const wrapper = shallow(<SprkRadio ariaDescribedBy={expected} />);
+    const wrapper = shallow(<SprkRadioItem ariaDescribedBy={expected} />);
     expect(
       wrapper.find('.sprk-b-Radio__input').prop('aria-describedby'),
     ).toEqual(expected);
@@ -53,7 +57,7 @@ describe('SprkRadio:', () => {
 
   it('should apply id', () => {
     const expected = 'testId';
-    const wrapper = shallow(<SprkRadio id={expected} />);
+    const wrapper = shallow(<SprkRadioItem id={expected} />);
     expect(wrapper.find('.sprk-b-Radio__label').prop('htmlFor')).toEqual(
       expected,
     );
@@ -62,7 +66,7 @@ describe('SprkRadio:', () => {
 
   it('should run the supplied onChangeFunc function for radioes', () => {
     const onRadioChangeMock = jest.fn();
-    const wrapper = mount(<SprkRadio onChangeFunc={onRadioChangeMock} />);
+    const wrapper = mount(<SprkRadioItem onChangeFunc={onRadioChangeMock} />);
     const radio = wrapper.find('input[type="radio"]');
     radio.simulate('change', { target: { value: 'test-value' } });
     expect(onRadioChangeMock.mock.calls.length).toBe(1);
