+++ react/src/spark-exports-react.js
@@ -11,8 +11,8 @@ export { default as SprkIcon } from './components/icons/SprkIcon';
 export { default as SprkTextInput } from './base/inputs/SprkTextInput/SprkTextInput';
 export { default as SprkRevealInput } from './base/inputs/SprkRevealInput/SprkRevealInput';
 export { default as SprkSelectionInput } from './base/inputs/SprkSelectionInput/SprkSelectionInput';
-export { default as SprkRadioGroup } from './base/inputs/SprkRadioGroup/SprkRadioGroup';
-export { default as SprkRadio } from './base/inputs/SprkRadio/SprkRadio';
+export { default as SprkRadioGroup } from './base/inputs/SprkRadio/SprkRadioGroup/SprkRadioGroup';
+export { default as SprkRadioItem } from './base/inputs/SprkRadio/SprkRadioItem/SprkRadioItem';
 export { default as SprkHelperText } from './base/inputs/SprkHelperText/SprkHelperText';
 export { default as SprkFieldset } from './base/inputs/SprkFieldset/SprkFieldset';
 export { default as SprkLegend } from './base/inputs/SprkLegend/SprkLegend';
